Disable Babel translation when running tests on the bots
Not only should this reduce the runtime of (in particular) the Windows bot, but it will also allow us to actually test the source code as-is rather than its Babel translated version.
Since the bots are running the current versions of Firefox/Chrome, testing ES6 code as-is should thus not be an issue.