Remove the font tests command, and related logic, from the bots
If https://github.com/mozilla/pdf.js/pull/17263 is in place, we will no
longer run the font tests on the bots, which means that the font tests
command should be removed. Moreover, we can get rid of all submodule
installations because the project won't have any submodules anymore
after that patch, which simplifies the code and reduces runtime of all
tests a bit (note that the submodules were also installed for tests
that didn't need them).