Use the locally-installed version of Gulp instead a globally-installed one
The globally-installed version of Gulp on the bots is old, and recently
we moved away from using globally-installed dependencies in favor of
using locally-installed ones for maintainability, isolation and
reproducibility of build environments. For more information, refer to
PRs https://github.com/mozilla/pdf.js/pull/17913 and
https://github.com/mozilla/pdf.js/pull/17489.

In https://github.com/mozilla/pdf.js/pull/18197 we have seen that even
though locally Gulp 5 is installed the global version is preferred by
the bots because the `gulp` command (available in the PATH) is used
instead of `npx gulp` which prefers the locally-installed version from
the `node_modules` folder.

This commit fixes the issue by making all Gulp invocations use `npx gulp`
instead, similar to what we already did in the GitHub Actions pipelines.